Rename Duke to Jarvis.
